@@ -1,7 +1,7 @@
 Linker::Application.routes.draw do
 
-  match 'ladmin/login' => 'ladmin#login'
-  match 'ladmin/logout' => 'ladmin#logout'
+  match 'ladmin/login' => 'ladmin#login', via: [:get, :post]
+  match 'ladmin/logout' => 'ladmin#logout', via: [:get, :post]
 
   resources :users
   resources :groups do
@@ -11,16 +11,17 @@ Linker::Application.routes.draw do
     end
   end
 
-  match 'search' => 'links#index'
-  match 'advanced_search' => 'links#advanced_search'
+  #match 'search' => 'links#index'
+  get 'search' => 'links#index'
+  get 'advanced_search' => 'links#advanced_search'
   resources :links do
     collection do
       get 'groups'
     end
   end
 
-  match 'toggle_full_details' => 'links#toggle_full_details'
-  match 'toggle_row_shading' => 'links#toggle_row_shading'
+  get 'toggle_full_details' => 'links#toggle_full_details'
+  get 'toggle_row_shading' => 'links#toggle_row_shading'
 
   get 'verify_link/:id', to: 'links#verify_link', as: :verify_link
   get 'unverify_link/:id', to: 'links#unverify_link', as: :unverify_link
