@@ -16,7 +16,7 @@ describe "Make sure last admin cannot be deleted" do
 
   let!(:user1) { User.create(username: 'user100', password: 'abc123', admin: false) }
 
-  its "how many" do
+  it "how many" do
     user1.destroy rescue
     expect(User.all.count).to eq 1
   end
@@ -28,7 +28,7 @@ describe "Make sure last admin cannot be deleted" do
   let!(:user1) { User.create(username: 'user100', password: 'abc123', admin: true) }
   let!(:user2) { User.create(username: 'user200', password: 'abc123', admin: true) }
 
-  its "how many" do
+  it "how many" do
     User.delete_one_admin(user1)
     User.delete_one_admin(user2) rescue # expect error
     expect(User.where(admin: true).count).to eq 1
